requirements.txt
pytest
selenium
webdriver-manager
python-dotenv
PyYAML
pytest-xdist
allure-pytest
